requirements.txt
flask
Flask-SQLAlchemy 
flask-marshmallow 
marshmallow-sqlalchemy 
marshmallow
werkzeug
pandas


